backend/requirements.txt
acres==0.5.0
adal==1.2.7
aiofiles==23.2.1
aistudio-sdk==0.3.8
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
bce-python-sdk==0.9.46
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.2
black==23.12.1
celery==5.3.6
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorclass==2.2.2
colorlog==6.9.0
compressed-rtf==1.0.7
configobj==5.0.9
configparser==7.2.0
coverage==7.10.7
cryptography==46.0.1
dnspython==2.8.0
doctr==1.8.0
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.1
email-validator==2.3.0
et-xmlfile==2.0.0
etelemetry==0.3.1
extract-msg==0.55.0
fastapi==0.115.4
filelock==3.19.1
fitz==0.0.1.dev2
flake8==6.1.0
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.35.1
idna==3.10
imagesize==1.4.1
imapclient==3.0.1
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
lark==1.1.9
libmagic==1.0
looseversion==1.3.0
lxml==6.0.2
mako==1.3.10
markupsafe==3.0.2
mccabe==0.7.0
modelscope==1.30.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
mypy==1.18.2
mypy-extensions==1.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
olefile==0.47
oletools==0.60.2
opencv-contrib-python==4.6.0.66
openpyxl==3.1.5
packaging==25.0
paddleocr
paddlex
pandas==2.1.4
passlib==1.7.4
pathlib==1.0.1
pathlib2==2.3.7
pathspec==0.12.1
pcodedmp==1.2.6
pdfminer-six==20221105
pdfplumber==0.10.4
pillow==10.4.0
pip==24.0
platformdirs==4.4.0
pluggy==1.6.0
prettytable==3.16.0
prompt-toolkit==3.0.52
prov==2.1.1
psutil==7.1.0
psycopg2-binary==2.9.10
puremagic==1.30
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycodestyle==2.11.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.11.0
pydot==4.0.1
pyflakes==3.1.0
pyjwt==2.10.1
pymupdf
pyparsing==3.2.5
pypdf2==3.0.1
pypdfium2==4.30.0
pytest==7.4.4
pytest-asyncio==0.21.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-msgraph==0.2.8
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
pyyaml==6.0.2
rdflib==7.2.1
red-black-tree-mod==1.22
redis==5.3.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
rtfde==0.1.2.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==2.7.0
setuptools==80.9.0
shapely==2.1.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.43
starlette==0.41.2
structlog==23.3.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.2.0
tqdm==4.67.1
traits==7.0.2
transformers==4.46.3
triton==2.2.0
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
